Refactor trivia to wrap into Suspense
